Start cleanup method
This will remove the window 'resize' event observer and disconnect the
resizeObserver.

A MutationObserver is also added for the attached canvas's parent so
this method is called whenever the canvas associated with a Niivue
instance is removed.

This cleans up references / addresses memory leaks, and e.g. prevents invalid
calls to this.resizeListener with a window resize after the associated
canvas has been removed from the DOM.